utils&bg-removal: Split fill processing initializationa and components
