+++ packages/plugin-utils/src/index.ts
@@ -5,3 +5,5 @@ export { default as FillProcessingMetadata } from './metadata/FillProcessingMeta
 export { default as processFill } from './processing/processFill';
 
 export { default as initializeFillProcessing } from './processing/initializeFillProcessing';
+
+export { default as registerFillProcessingComponents } from './processing/registerFillProcessingComponents';
